requirements.txt
accelerate==0.23.0
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
compel==2.0.2
contextlib2==21.6.0
controlnet-aux==0.0.7
diffusers==0.21.2
einops==0.6.1
filelock==3.12.4
fsspec==2023.9.2
huggingface-hub==0.17.3
idna==3.4
imageio==2.31.4
importlib-metadata==6.8.0
Jinja2==3.1.2
lazy_loader==0.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
omegaconf==2.3.0
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.1
psutil==5.9.5
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.1
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-image==0.21.0
scipy==1.11.3
sympy==1.12
tifffile==2023.9.26
timm==0.9.7
tokenizers==0.13.3
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.0+cu118
--extra-index-url https://download.pytorch.org/whl/cu118
torchaudio==2.1.0+cu118
--extra-index-url https://download.pytorch.org/whl/cu118
torchvision==0.16.0+cu118
tqdm==4.66.1
transformers==4.33.2
typing_extensions==4.8.0
urllib3==2.0.5
zipp==3.17.0
